third_party/blink/web_tests/editing/pasteboard/select-element-1-expected.txt

This tests copy/paste of select elements. All the options should be included.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("select") === null is true
PASS select.options[0].value is "One"
PASS select.options[1].value is "Two"
PASS select.options[2].value is "Three"
PASS successfullyParsed is true

TEST COMPLETE
